requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
idna==3.4
iniconfig==2.0.0
kombu==5.3.1
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.38
pytest==7.4.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
